+++ src/CustomizeGeyser/_FumiKMod.cs
@@ -1,0 +1,18 @@
+﻿using HarmonyLib;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace CustomizeGeyser
+{
+    public class FumiKMod : KMod.UserMod2
+    {
+        public override void OnLoad(Harmony harmony)
+        {
+            CustomizeGeyserState.OnLoad();
+            base.OnLoad(harmony);
+        }
+    }
+}